+++ routes/appPreferencesRoutes.js
@@ -2,45 +2,10 @@ const express = require('express');
 const router = express.Router();
 const { User } = require('../models/shared/UserSchema');
 const { authenticateToken } = require('../middleware/authMiddleware');
+const { findUserFromRequest, buildUserFilter } = require('../utils/userLookup');
 
 // Current schema version (YYYYMMDDXXX format)
 const CURRENT_SCHEMA_VERSION = 20250812001;
-
-/**
- * Find user by req.user (supports email OR provider-based auth)
- * Works with Twitter, Nostr, and email-based users
- */
-async function findUserFromRequest(req, selectFields = '') {
-  if (req.user.email) {
-    return User.findOne({ email: req.user.email }).select(selectFields);
-  } else if (req.user.provider && req.user.providerId) {
-    return User.findOne({
-      'authProvider.provider': req.user.provider,
-      'authProvider.providerId': req.user.providerId
-    }).select(selectFields);
-  } else if (req.user.id) {
-    // Fallback to user ID if available
-    return User.findById(req.user.id).select(selectFields);
-  }
-  return null;
-}
-
-/**
- * Build query filter from req.user (supports email OR provider-based auth)
- */
-function buildUserFilter(req) {
-  if (req.user.email) {
-    return { email: req.user.email };
-  } else if (req.user.provider && req.user.providerId) {
-    return {
-      'authProvider.provider': req.user.provider,
-      'authProvider.providerId': req.user.providerId
-    };
-  } else if (req.user.id) {
-    return { _id: req.user.id };
-  }
-  return null;
-}
 
 // Deep merge helper that merges arrays of objects by "id" when present,
 // otherwise replaces arrays. Objects are merged recursively.
